feat: update StatsDto to use Long for hits and add date range validation in StatsServiceImpl

+++ ewm-stats-service/stats-dto/src/main/java/ru/practicum/ewm/stats/dto/response/StatsDto.java
@@ -6,9 +6,9 @@ import lombok.Data;
 public class StatsDto {
     private String app;
     private String uri;
-    private Integer hits;
+    private Long hits;
 
-    public StatsDto(String app, String uri, Integer hits) {
+    public StatsDto(String app, String uri, Long hits) {
         this.app = app;
         this.uri = uri;
         this.hits = hits;
